+++ changelog.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - removed duplicate packages between modules to prepare for java modules
 - caught Jackson configuration exception to improve resilience with other Jackson versions in classpath
 - moved Junit4 to separate module to reduce size of jar-with-dependencies, simplify code and increase build speed
+- enabled case insensitive matching for regex matches
 
 ### Fixed
 - fixed error where ClientAndServer does fully wait for client to stop